chore: complimentary pr for CD migration changes (#35876)
## Description

> CE changes for CD migration on EE pr:
https://github.com/appsmithorg/appsmith-ee/pull/4927

## Automation

/ok-to-test tags="@tag.Git"

### :mag: Cypress test results
